+++ go.mod
@@ -5,7 +5,7 @@ go 1.16
 require (
 	github.com/Microsoft/go-winio v0.5.2 // indirect
 	github.com/docker/distribution v2.8.1+incompatible // indirect
-	github.com/docker/docker v20.10.16+incompatible
+	github.com/docker/docker v20.10.18+incompatible
 	github.com/docker/go-connections v0.4.0 // indirect
 	github.com/docker/go-units v0.4.0 // indirect
 	github.com/gogo/protobuf v1.3.2 // indirect
@@ -16,5 +16,6 @@ require (
 	github.com/procyon-projects/chrono v1.0.0
 	github.com/sirupsen/logrus v1.8.1 // indirect
 	golang.org/x/time v0.0.0-20220411224347-583f2d630306 // indirect
+	golang.org/x/vuln v0.0.0-20221004204012-7ee453d0a82a
 	gotest.tools/v3 v3.2.0 // indirect
 )
